+++ CHANGELOG.md
@@ -4,8 +4,13 @@ DemonTech - Technology and a lick of flame.
 ## [Unreleased]
 
 ### Added
+#### Modules
  - `libraries/datagen` - Data generation utilities. 
  - `libraries/sync` - KSP plugin to generate BlockEntity data helpers.
+
+#### Items
+ - Thermal Pad
+ - Adaptive Thermal Pad
 
 ### Changed